requirements.txt
lxml==3.7.0
python-dateutil==2.6.0
requests==2.12.4
